requirements.txt
numpy==2.2.3
numpy-stl==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyqtgraph==0.13.3
python-utils==3.9.1
typing_extensions==4.12.2
uv==0.6.2
